+++ lib/ui/navigation/main_navigation.dart
@@ -31,6 +31,7 @@ abstract class MainNavigationRouteNames {
   static const favoriteTvs = '/main_screen/favorite_tvs';
   static const peopleDetails = '/main_screen/peopleDetails';
   static const movieDetailsFullCastAndCrewList = '/main_screen/movie_details/full_cast_and_crew';
+  static const peopleDetailsKnownForList = '/main_screen/people_details/full_known_for';
   static const settings = '/personal_widget';
   static const networkConnectionError = '/errors/network_connection';
 }
@@ -67,6 +68,12 @@ class MainNavigation {
         return MaterialPageRoute(
           builder: (_) => _screenFactory.makeFullCastAndCrewList(movieId),
         );
+      case MainNavigationRouteNames.peopleDetailsKnownForList:
+        final arguments = settings.arguments;
+        final peopleId = arguments is int ? arguments : 0;
+        return MaterialPageRoute(
+          builder: (_) => _screenFactory.makePeopleDetailsKnownForList(peopleId),
+        );
       case MainNavigationRouteNames.movieTrailer:
         final arguments = settings.arguments;
         final youtubeKey = arguments is String ? arguments : '';
@@ -90,7 +97,7 @@ class MainNavigation {
         final arguments = settings.arguments;
         final youtubeKey = arguments is String ? arguments : '';
         return MaterialPageRoute(
-          builder: (context) => TvTrailerWidget(tvYoutubeKey: youtubeKey),
+          builder: (context) => TvDetailsTrailerWidget(tvYoutubeKey: youtubeKey),
         );
       default:
         const widget = Text('Navigation error ');