+++ lib/ui/widgets/movie_details/components/cast_and_crew.dart
@@ -64,7 +64,7 @@ class _MovieActorListWidget extends StatelessWidget {
     var actorsData = context.watch<MovieDetailsCubit>().data.actorsData;
     if (actorsData.isEmpty) return const SizedBox.shrink();
     return ListView.builder(
-      itemCount: 9,
+      itemCount: 8,
       itemExtent: 120,
       scrollDirection: Axis.horizontal,
       itemBuilder: (BuildContext context, int index) {