+++ lib/ui/widgets/movie_details/components/movie_details_all_videos_data.dart
@@ -1,0 +1,21 @@
+class MovieDetailsAllVideosData {
+  final String name;
+  final String key;
+  final String site;
+  final int size;
+  final String type;
+  final bool official;
+  final String publishedAt;
+  final String id;
+
+  const MovieDetailsAllVideosData({
+    required this.name,
+    required this.key,
+    required this.site,
+    required this.size,
+    required this.type,
+    required this.official,
+    required this.publishedAt,
+    required this.id,
+  });
+}