+++ lib/ui/widgets/movie_details/movie_details_state.dart
@@ -21,7 +21,8 @@ class MovieDetailsCubitState {
   final bool isFavorite;
   final List<MovieDetailsRecommendationsData> recommendations;
   final List<MovieDetailsVideosData> videos;
-  // final MovieDetailsExternalIdsData externalIds;
+  final List<MovieDetailsAllVideosData> allVideos;
+  // final List<MovieDetailsExternalIdsData> externalIds;
   // final List<MovieDetailsReviewsData> reviews;
   // final List<MovieDetailsSimilarData> similar;
   // final List<BelongsToCollectionData>? collection;
@@ -46,6 +47,8 @@ class MovieDetailsCubitState {
     required this.isFavorite,
     required this.recommendations,
     required this.videos,
+    required this.allVideos,
+    // required this.externalIds,
     // required this.externalIds,
     // required this.reviews,
     // required this.similar,
@@ -75,10 +78,11 @@ class MovieDetailsCubitState {
           isLoading == other.isLoading &&
           isFavorite == other.isFavorite &&
           recommendations == other.recommendations &&
-          // externalIds == other.externalIds;
+          // externalIds == other.externalIds &&
           // reviews == other.reviews &&
           // similar == other.similar;
-          videos == other.videos;
+          videos == other.videos &&
+          allVideos == other.allVideos;
           // collection == other.collection;
 
   @override
@@ -101,10 +105,12 @@ class MovieDetailsCubitState {
       isLoading.hashCode ^
       isFavorite.hashCode ^
       recommendations.hashCode ^
+      // externalIds.hashCode ^
       // externalIds.hashCode;
       // reviews.hashCode ^
       // similar.hashCode;
-      videos.hashCode;
+      videos.hashCode ^
+      allVideos.hashCode;
       // collection.hashCode;
 
   MovieDetailsCubitState copyWith({
@@ -127,7 +133,8 @@ class MovieDetailsCubitState {
     bool? isFavorite,
     List<MovieDetailsRecommendationsData>? recommendations,
     List<MovieDetailsVideosData>? videos,
-    // MovieDetailsExternalIdsData? externalIds,
+    List<MovieDetailsAllVideosData>? allVideos,
+    // List<MovieDetailsExternalIdsData>? externalIds,
     // List<MovieDetailsReviewsData>? reviews,
     // List<MovieDetailsSimilarData>? similar,
     // List<BelongsToCollectionData>? collection,
@@ -152,6 +159,7 @@ class MovieDetailsCubitState {
       isFavorite: isFavorite ?? this.isFavorite,
       recommendations: recommendations ?? this.recommendations,
       videos: videos ?? this.videos,
+      allVideos: allVideos ?? this.allVideos,
       // externalIds: externalIds ?? this.externalIds,
       // reviews: reviews ?? this.reviews,
       // similar: similar ?? this.similar,