+++ lib/ui/widgets/movie_details/movie_details_widget.dart
@@ -1,5 +1,4 @@
 // Flutter imports:
-import 'package:comics_db_app/ui/widgets/movie_details/components/collection_widget.dart';
 import 'package:flutter/material.dart';
 
 // Package imports:
@@ -11,7 +10,7 @@ import 'package:comics_db_app/ui/components/custom_floating_action_button.dart';
 import 'package:comics_db_app/ui/widgets/movie_details/components/cast_and_crew.dart';
 import 'package:comics_db_app/ui/widgets/movie_details/components/description_widget.dart';
 import 'package:comics_db_app/ui/widgets/movie_details/components/peoples_widget.dart';
-import 'package:comics_db_app/ui/widgets/movie_details/components/recommendations_widget.dart';
+import 'package:comics_db_app/ui/widgets/movie_details/components/movie_details_recommendations_widget.dart';
 import 'package:comics_db_app/ui/widgets/movie_details/components/top_poster_widget.dart';
 import 'package:comics_db_app/ui/widgets/movie_details/movie_details_cubit.dart';
 
@@ -35,7 +34,6 @@ class _MovieDetailsWidgetState extends State<MovieDetailsWidget> {
   Widget build(BuildContext context) {
     var cubit = context.watch<MovieDetailsCubit>();
     final favorite = cubit.state.isFavorite;
-    // final String trailerKey = cubit.state.videos.first.key;
 
     return Scaffold(
       appBar: const CustomDetailsAppBar(title: ''),
@@ -44,12 +42,13 @@ class _MovieDetailsWidgetState extends State<MovieDetailsWidget> {
         children: [
           Column(
             children: const [
-              MovieTopPosterWidget(),
-              PeoplesWidget(),
-              DescriptionWidget(),
-              CastWidget(),
+               MovieTopPosterWidget(),
+               PeoplesWidget(),
+               DescriptionWidget(),
+               CastWidget(),
+                  // MovieDetailsAllVideosWidget(),
               // const MovieDetailsReviewsWidget(),
-              MovieDetailsRecommendations(),
+               MovieDetailsRecommendations(),
               // const MovieDetailsSimilarWidget(),
               // const MovieSimilarWidget(),
             ],