+++ lib/ui/widgets/people_details/components/character_data.dart
@@ -1,10 +1,12 @@
 class PeopleDetailsCharacterData {
+  final int id;
   final String character;
   final String title;
   final String? posterPath;
   final String? backdropPath;
 
   PeopleDetailsCharacterData({
+    required this.id,
     required this.character,
     required this.title,
     required this.posterPath,