+++ lib/ui/widgets/people_details/components/people_details_data.dart
@@ -16,7 +16,6 @@ class PeopleDetailsData {
   bool adult = true;
   String imdbId = '';
   String? homepage = '';
-  List<PeopleDetailsCharacterData> charactersData =
-      const <PeopleDetailsCharacterData>[];
-  // List<KnownForData> knownFor = const <KnownForData>[];
+  List<PeopleDetailsCharacterData> charactersData = const <PeopleDetailsCharacterData>[];
+// List<KnownForData> knownFor = const <KnownForData>[];
 }