+++ lib/ui/widgets/tv_details/components/tv_details_recommendations_data.dart
@@ -1,0 +1,13 @@
+class TvDetailsRecommendationsData {
+  final int id;
+  final String? posterPath;
+  final String? backdropPath;
+  final String name;
+
+  const TvDetailsRecommendationsData({
+    required this.id,
+    this.posterPath,
+    this.backdropPath,
+    required this.name,
+  });
+}