+++ lib/ui/widgets/tv_details/components/tv_details_recommendations_widget.dart
@@ -1,4 +1,7 @@
 // Flutter imports:
+import 'package:comics_db_app/core/dark_theme_colors.dart';
+import 'package:comics_db_app/domain/blocs/theme/theme_bloc.dart';
+import 'package:comics_db_app/ui/widgets/tv_details/tv_details_cubit.dart';
 import 'package:flutter/material.dart';
 
 // Package imports:
@@ -13,12 +16,12 @@ import 'package:comics_db_app/ui/components/loading_indicator_widget.dart';
 import 'package:comics_db_app/ui/navigation/main_navigation.dart';
 import 'package:comics_db_app/ui/widgets/movie_details/movie_details_cubit.dart';
 
-class MovieDetailsRecommendations extends StatelessWidget {
-  const MovieDetailsRecommendations({Key? key}) : super(key: key);
+class TvDetailsRecommendationsWidget extends StatelessWidget {
+  const TvDetailsRecommendationsWidget({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
-    var recommendationsData = context.watch<MovieDetailsCubit>().data.recommendationsData;
+    var recommendationsData = context.watch<TvDetailsCubit>().data.recommendationsData;
     if (recommendationsData.isEmpty) return const SizedBox.shrink();
     return Padding(
       padding: const EdgeInsets.only(top: 20.0, left: 20.0, right: 20.0),
@@ -30,19 +33,19 @@ class MovieDetailsRecommendations extends StatelessWidget {
             'Recommendations',
             style: Theme.of(context).textTheme.titleMedium,
           ),
-          const _MovieDetailsRecommendationsWidget(),
+          const _TvDetailsRecommendationsListWidget(),
         ],
       ),
     );
   }
 }
 
-class _MovieDetailsRecommendationsWidget extends StatelessWidget {
-  const _MovieDetailsRecommendationsWidget({Key? key}) : super(key: key);
+class _TvDetailsRecommendationsListWidget extends StatelessWidget {
+  const _TvDetailsRecommendationsListWidget({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
-    var recommendationsData = context.watch<MovieDetailsCubit>().data.recommendationsData;
+    var recommendationsData = context.watch<TvDetailsCubit>().data.recommendationsData;
     if (recommendationsData.isEmpty) return const SizedBox.shrink();
     return Padding(
       padding: const EdgeInsets.only(bottom: 12.0),
@@ -59,12 +62,23 @@ class _MovieDetailsRecommendationsWidget extends StatelessWidget {
                 child: Container(
                   width: 220,
                   clipBehavior: Clip.antiAlias,
-                  decoration: const BoxDecoration(
-                    borderRadius: BorderRadius.all(
-                      Radius.circular(25.0),
+                  decoration: BoxDecoration(
+                    color: context.read<ThemeBloc>().isDarkTheme ? DarkThemeColors.kPrimaryColor : Colors.white,
+                    border: Border.all(
+                      color: context.read<ThemeBloc>().isDarkTheme ? Colors.white.withOpacity(0.2) : Colors.black.withOpacity(0.2),
+                    ),
+                    boxShadow: [
+                      BoxShadow(
+                        color: Colors.purple.withOpacity(0.1),
+                        blurRadius: 8,
+                        offset: const Offset(0, 2),
+                      ),
+                    ],
+                    borderRadius: const BorderRadius.all(
+                      Radius.circular(10.0),
                     ),
                   ),
-                  child: MovieDetailsItemRecommendationsWidget(index: index),
+                  child: TvDetailsItemRecommendationsWidget(index: index),
                 ),
               );
             },
@@ -75,21 +89,21 @@ class _MovieDetailsRecommendationsWidget extends StatelessWidget {
   }
 }
 
-class MovieDetailsItemRecommendationsWidget extends StatelessWidget {
+class TvDetailsItemRecommendationsWidget extends StatelessWidget {
   final int index;
 
-  const MovieDetailsItemRecommendationsWidget({
+  const TvDetailsItemRecommendationsWidget({
     super.key,
     required this.index,
   });
 
   @override
   Widget build(BuildContext context) {
-    final cubit = context.read<MovieDetailsCubit>();
+    final cubit = context.read<TvDetailsCubit>();
     final recommendationsData = cubit.data.recommendationsData[index];
     final recommendationsId = recommendationsData.id;
     final backdropPath = recommendationsData.backdropPath;
-    final title = recommendationsData.title;
+    final name = recommendationsData.name;
     return InkWell(
       onTap: () => Navigator.of(context).pushNamed(MainNavigationRouteNames.movieDetails, arguments: recommendationsId),
       child: ClipRRect(
@@ -97,10 +111,10 @@ class MovieDetailsItemRecommendationsWidget extends StatelessWidget {
           children: [
             backdropPath != null
                 ? CachedNetworkImage(
-                    imageUrl: ImageDownloader.imageUrl(backdropPath),
-                    placeholder: (context, url) => const LoadingIndicatorWidget(),
-                    errorWidget: (context, url, dynamic error) => Image.asset(AppImages.imageNotAvailableHorizontal),
-                  )
+              imageUrl: ImageDownloader.imageUrl(backdropPath),
+              placeholder: (context, url) => const LoadingIndicatorWidget(),
+              errorWidget: (context, url, dynamic error) => Image.asset(AppImages.imageNotAvailableHorizontal),
+            )
                 : Image.asset(AppImages.imageNotAvailableHorizontal),
             Expanded(
               child: Padding(
@@ -108,7 +122,7 @@ class MovieDetailsItemRecommendationsWidget extends StatelessWidget {
                 child: Column(
                   crossAxisAlignment: CrossAxisAlignment.start,
                   children: [
-                    CustomCastListTextWidget(text: title, maxLines: 1),
+                    CustomCastListTextWidget(text: name, maxLines: 1),
                   ],
                 ),
               ),