+++ lib/ui/widgets/tv_trailer/tv_trailer_widget.dart
@@ -4,16 +4,17 @@ import 'package:flutter/material.dart';
 // Package imports:
 import 'package:youtube_player_flutter/youtube_player_flutter.dart';
 
-class TvTrailerWidget extends StatefulWidget {
-  final String tvYoutubeKey;
+// TODO analog in movie details widget
+class TvDetailsTrailerWidget extends StatefulWidget {
+  final String? tvYoutubeKey;
 
-  const TvTrailerWidget({Key? key, required this.tvYoutubeKey}) : super(key: key);
+  const TvDetailsTrailerWidget({Key? key, required this.tvYoutubeKey}) : super(key: key);
 
   @override
-  State<TvTrailerWidget> createState() => _TvTrailerWidgetState();
+  State<TvDetailsTrailerWidget> createState() => _TvDetailsTrailerWidgetState();
 }
 
-class _TvTrailerWidgetState extends State<TvTrailerWidget> {
+class _TvDetailsTrailerWidgetState extends State<TvDetailsTrailerWidget> {
   late final YoutubePlayerController _controller;
 
   @override
@@ -21,7 +22,7 @@ class _TvTrailerWidgetState extends State<TvTrailerWidget> {
     super.initState();
 
     _controller = YoutubePlayerController(
-      initialVideoId: widget.tvYoutubeKey,
+      initialVideoId: widget.tvYoutubeKey ?? '',
       flags: const YoutubePlayerFlags(
         autoPlay: true,
         mute: true,
@@ -31,21 +32,26 @@ class _TvTrailerWidgetState extends State<TvTrailerWidget> {
 
   @override
   Widget build(BuildContext context) {
-    return YoutubePlayerBuilder(
-      player: YoutubePlayer(
-        controller: _controller,
-        showVideoProgressIndicator: true,
-      ),
-      builder: (context, player) {
-        return Scaffold(
-          appBar: AppBar(),
-          body: Column(
-            children: [
-              player,
-            ],
+    return Padding(
+      padding: const EdgeInsets.only(left: 20.0, right: 20.0, top: 20.0),
+      child: Column(
+        crossAxisAlignment: CrossAxisAlignment.start,
+        children: [
+          YoutubePlayerBuilder(
+            player: YoutubePlayer(
+              controller: _controller,
+              showVideoProgressIndicator: true,
+            ),
+            builder: (context, player) {
+              return Column(
+                children: [
+                  player,
+                ],
+              );
+            },
           ),
-        );
-      },
+        ],
+      ),
     );
   }
 }